requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
reportlab==4.0.0
Pillow==10.0.1
Flask-Login==0.6.2
